BACKPORT: sched/x86: Rewrite the switch_to() code
Move the low-level context switch code to an out-of-line asm stub instead of
using complex inline asm.  This allows constructing a new stack frame for the
child process to make it seamlessly flow to ret_from_fork without an extra
test and branch in __switch_to().  It also improves code generation for
__schedule() by using the C calling convention instead of clobbering all
registers.

BUG=chromium:836330
TEST=Build and boot

Backport Notes:
- the original switch_to implementation inside switch_to.h
had been modified to include retpoline related code in the v4.4 commit
18bb117d1b769.
When moving switch_to to entry_32.S and entry_64.S, the retpoline code
needs to be moved as well; similar to how it is done for the v4.14
kernel.
- In the original upstream patch, inside arch/x86/include/asm/switch_to.h,
switch_to expected a call to prepare_switch_to. This was introduced in
e37e43a49("x86/mm/64: Enable vmapped stacks (CONFIG_HAVE_ARCH_VMAP_STACK=y)").
As VMAP_STACK is not being backported, that patch is not pulled in, and
the call to prepare_switch_to is not present.

Change-Id: I127d75f5312d52892b8256609e98902a89e948c6
Signed-off-by: Brian Gerst <brgerst@gmail.com>
Reviewed-by: Josh Poimboeuf <jpoimboe@redhat.com>
Cc: Andy Lutomirski <luto@kernel.org>
Cc: Borislav Petkov <bp@alien8.de>
Cc: Denys Vlasenko <dvlasenk@redhat.com>
Cc: H. Peter Anvin <hpa@zytor.com>
Cc: Linus Torvalds <torvalds@linux-foundation.org>
Cc: Peter Zijlstra <peterz@infradead.org>
Cc: Thomas Gleixner <tglx@linutronix.de>
Link: http://lkml.kernel.org/r/1471106302-10159-5-git-send-email-brgerst@gmail.com
Signed-off-by: Ingo Molnar <mingo@kernel.org>
(cherry picked from commit 0100301bfdf56a2a370c7157b5ab0fbf9313e1cd)
Signed-off-by: Zubin Mithra <zsm@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1077206
Reviewed-by: Guenter Roeck <groeck@chromium.org>