BACKPORT: drm/i915: Introduce i915_ggtt_offset()
This little helper only exists to safely discard the upper unused 32bits
of the general 64-bit VMA address - as we know that all Global GTT
currently are less than 4GiB in size and so that the upper bits must be
zero. In many places, we use a u32 for the global GTT offset and we want
to document where we are discarding the full VMA offset.

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>
Reviewed-by: Joonas Lahtinen <joonas.lahtinen@linux.intel.com>
Link: http://patchwork.freedesktop.org/patch/msgid/1471254551-25805-28-git-send-email-chris@chris-wilson.co.uk
(cherry picked from commit bde13ebdab0778b758b267ff9e38d6c10a42bdc3)
from drm-intel-next-2016-08-22
Signed-off-by: Tarun <tarun.vyas@intel.com>

BUG=b:36446508
TEST=No regressions in KBL and APL based devices

 Conflicts: [Tarun: Minor ctx delta]
	drivers/gpu/drm/i915/intel_lrc.c

Change-Id: If41d07b4d5c10d7a0166086b4dc91f4011ba08ba