UPSTREAM: drm/i915: Bump the inactive tracking for all VMA accessed
We track the LRU access for eviction and bump the last access for the
user GGTT on set-to-gtt. When we do so we need to not only bump the
primary GGTT VMA but all partials as well. Similarly we want to
bump the last access tracking for when unpinning an object from the
scanout so that they do not get promptly evicted and hopefully remain
available for reuse on the next frame.

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>
Reviewed-by: Joonas Lahtinen <joonas.lahtinen@linux.intel.com>
Link: http://patchwork.freedesktop.org/patch/msgid/20160818161718.27187-29-chris@chris-wilson.co.uk
(cherry picked from commit 383d5823e938129fe10c76e0032a22734849049c)
from drm-intel-next-2016-08-22
Signed-off-by: Tarun <tarun.vyas@intel.com>

BUG=b:36446508
TEST=No regressions in KBL and APL based devices

Change-Id: I9826c3cfbdf203cfb10b57deb3968d780690a070
Reviewed-on: https://chromium-review.googlesource.com/460642
Commit-Ready: Sean Paul <seanpaul@chromium.org>
Tested-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Sean Paul <seanpaul@chromium.org>