ANDROID: FIXUP: sched/tune: update accouting before CPU capacity
The SchedTune tasks accounting is used to identify how many tasks are in
a boostgroup and thus to bias the selection of an OPP based on the
maximum boost value of the active boostgroups.
The current implementation however update the accounting after CPU
capacity has been update. This has two effects:
a) when we enqueue a boosted task, we do not immediately boost its CPU
b) when we dequeue a boosted task, we can keep a CPU boosted even if not
   required

This patch change the order of the SchedTune accounting and SchedFreq
updated to ensure to have always an updated representation of which
boosted tasks are runnable on a CPU before updating its capacity.

BUG=chrome-os-partner:58667
TEST=build/boot on kevin

Change-Id: I92c9136c0d55e7621f086de8ec1d75470997f06c
Reported-by: Leo Yan <leo.yan@linaro.org>
Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>
(cherry picked from commit a85045c03470e88923ed98bb5917b289fa75203
 from android-msm/android-4.4.y)
Signed-off-by: Sonny Rao <sonnyrao@chromium.org>
Previous-Reviewed-on: https://chromium-review.googlesource.com/420089
(cherry picked from commit 0d94db15c9e54f1c4579b83a3c8c5e7122aad0f6)
Reviewed-on: https://chromium-review.googlesource.com/424116